deep_fake/requirements.txt
Flask==2.0.1
tensorflow==2.11.0
opencv-python==4.5.5.64
pymongo==4.10.1
dnspython==2.3.0
python-dotenv==0.19.0
Werkzeug==2.0.1
numpy==1.22.3